Subject handles unknown OIDs even better.
OpenSSL::X509::Name throws away unknown OIDs, replacing them with UNDEF,
in its array representation. This is unacceptably broken, so in lieu of
waiting for it to be fixed in Ruby, we now sanitize (with
R509::NameSanitizer) the OpenSSL::X509::Name and replace the UNDEF with
the unknown OID that it's supposed to be. In other words, this is
awesome.