modify the extensions to allow building as well as parsing (start #38)
This is essentially a copy and slight refactor of the code that resides
in R509::CertificateAuthority::Signer#build_extensions. Tests have not
been ported and yardoc is non-existent. I'm not 100% sure this is the
direction I want to go, but it's moderately promising.

With this r509 can potentially allow you to pass an array of pre-built
extensions to the signer rather than having the signer parse the config
and profiles and build them itself.

The biggest problem with the current setup is that each object can take
very different arguments in the constructor. For example, to build an
OCSPNoCheck extension you just do

R509::Cert::Extensions::OCSPNoCheck.new

and SAN:

R509::Cert::Extensions::SubjectAlternativeName.new(['cnn.com','127.0.0.1'])

and subjectKeyIdentifier

R509::Cert::Extensions::SubjectKeyIdentifier.new(:public_key =>
subject_public_key)

and AuthorityKeyIdentifier

R509::Cert::Extensions::AuthorityKeyIdentifier.new(
  :issuer_certificate => r509_cert_obj,
  :value => "keyid:always")