refactor Signer and tests (more massive refactoring for #38)
This refactor completely changes the behavior of the Signer. It now
takes a limited # of options in the hash and assumes you've built your
extensions prior to calling the signer. You can either build them
manually or use the ProfileEnforcer to create them.

Also included are some README updates.

Test coverage around the extension creation will be improved next